backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
pyserial
SpeechRecognition
pyaudio
pynput
openai
